requirements.txt
analytics-python==1.4.0
appdirs==1.4.4
argcomplete==2.0.0
async-generator==1.10
attrs==21.4.0
azure-core==1.22.1
azure-storage-blob==12.11.0
backoff==1.10.0
bcrypt==3.2.0
black==22.3.0
boto3==1.21.43
botocore==1.24.43
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cycler==0.11.0
determined==0.17.14
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
filelock==3.6.0
flake8==4.0.1
fonttools==4.32.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
jmespath==1.0.0
jsonschema==3.2.0
kiwisolver==1.4.2
lmdb==1.3.0
lomond==0.3.3
matplotlib==3.5.1
mccabe==0.6.1
monotonic==1.6
msrest==0.6.21
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
paramiko==2.10.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
websocket-client==0.59.0
websockets==10.3
yogadl==0.1.4
